(chore): 빌드 수정
