+++ app/api/auth/service/service.go
@@ -2,8 +2,8 @@ package service
 
 import (
 	"main/app/api/auth/resource"
-	"main/common/util"
 	"main/config"
+	"main/constant/common"
 	"main/database/repository"
 	"time"
 
@@ -184,7 +184,7 @@ func (as *authService) FindVisual(userId uint) (res *resource.FindVisualResponse
 	var collectRate uint
 
 	// 2. 수집률로 조건식을 사용하여 코드분류
-	code := util.PercentCal(collectRate)
+	code := common.PercentCal(collectRate)
 
 	// 3.  수집률을 담아 만들어놓은 레포지토리를 사용해서 데이터를 가져온다
 	visualFind, err := userRepository.FindVisual(code)
@@ -213,7 +213,7 @@ func (as *authService) FindVisualCode(userId uint) (res *resource.FindVisualCode
 	var collectRate uint
 
 	// 2. 수집률로 조건식을 사용하여 코드분류
-	code := util.PercentCal(collectRate)
+	code := common.PercentCal(collectRate)
 
 	// 3.  수집률을 담아 만들어놓은 레포지토리를 사용해서 데이터를 가져온다
 	visualFind, err := userRepository.FindVisualCode(code)
