+++ config/constant.go
@@ -13,4 +13,8 @@ const (
 	DATABASE_RW_PASSWORD = "database.rw.password"
 	DATABASE_RO_USER     = "database.ro.user"
 	DATABASE_RO_PASSWORD = "database.ro.password"
+	GRPC_HISTORY_PORT    = "grpc.history.port"
+	GRPC_HISTORY_HOST    = "grpc.history.host"
+	GRPC_MAP_PORT        = "grpc.map.port"
+	GRPC_MAP_HOST        = "grpc.map.host"
 )
