--- database/entity/patient.go
@@ -1,10 +1,0 @@
-package entity
-
-import "gorm.io/gorm"
-
-type Patient struct {
-	gorm.Model
-	Name   string
-	Age    int
-	Gender string
-}
